requirements.txt
amqp==2.3.2
anyjson==0.3.3
asgiref==3.2.10
beautifulsoup4==4.6.0
billiard==3.5.0.5
bs4==0.0.1
celery==4.2.1
certifi==2018.11.29
cffi==1.14.1
chardet==3.0.4
cloudinary==1.20.0
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.5
django-admin-tools==0.8.1
django-allauth==0.41.0
django-celery-results==1.0.1
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-heroku==0.2.0
django-import-export==2.4.0
django-ipware==2.1.0
django-paypal==1.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-simple-history==2.9.0
django-timezone-field==4.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
ez-setup==0.9
gunicorn==19.7.1
idna==2.7
jdcal==1.4.1
kiwisolver==1.2.0
kombu==4.2.1
lxml==4.5.2
MarkupPy==1.14
matplotlib==3.3.2
mws==0.8.6
mysqlclient==2.0.1
numpy==1.19.2
oauthlib==2.0.6
odfpy==1.4.1
openpyxl==3.0.4
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
Pillow==8.0.0
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
pymysql==0.10.1
pyparsing==2.4.7
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.1
python-memcached==1.59
python3-openid==3.1.0
pytz==2018.7
PyYAML==5.3.1
requests==2.20.1
requests-oauthlib==0.8.0
selenium==3.141.0
simplejson==3.13.2
six==1.15.0
sqlparse==0.2.4
stripe==1.79.0
tablib==2.0.0
twilio==6.46.0
urllib3==1.24.1
vine==1.1.4
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
